Refactor network inspector into feature modules
