+++ snapo-desktop-electron/src/features/network-inspector/lib/exportActions.ts
@@ -1,0 +1,46 @@
+import type { NetworkClient } from "../../../network/client";
+import { applyRequestBodies, type InspectorRecord, type RequestRecord } from "../../../network/cdp";
+import { buildHar, harFileName, makeCurlCommand } from "../../../network/exporters";
+
+export async function copyCurl(client: NetworkClient, request: RequestRecord): Promise<void> {
+  let hydrated = request;
+  if (request.requestBody == null) {
+    try {
+      hydrated = applyRequestBodies(
+        request,
+        await client.loadBodies({
+          deviceId: request.server.deviceId,
+          socketName: request.server.socketName,
+          requestId: request.requestId
+        })
+      );
+    } catch {
+      hydrated = request;
+    }
+  }
+  await navigator.clipboard.writeText(makeCurlCommand(hydrated));
+}
+
+export async function exportAsHar(client: NetworkClient, records: InspectorRecord[]): Promise<void> {
+  if (records.length === 0) return;
+  const hydrated = await Promise.all(
+    records.map(async (record) => {
+      if (record.kind !== "request" || (record.requestBody != null && record.responseBody != null)) return record;
+      try {
+        const bodies = await client.loadBodies({
+          deviceId: record.server.deviceId,
+          socketName: record.server.socketName,
+          requestId: record.requestId
+        });
+        return applyRequestBodies(record, bodies);
+      } catch {
+        return record;
+      }
+    })
+  );
+  await client.saveFile({
+    defaultPath: harFileName(hydrated.length),
+    data: buildHar(hydrated),
+    mimeType: "application/har+json"
+  });
+}
